added Language enum, *css

+++ src/main/java/pl/masters/coding/teacher/TeacherServiceImpl.java
@@ -16,11 +16,10 @@ public class TeacherServiceImpl  {
         return teacherRepository.findAll();
     }
 
-    public void deleteTeacher(Long id) {
+    public void deleteTeacher(Integer id) {
         List<Teacher> teacherList = teacherRepository.findAll();
         teacherList.removeIf(teacher -> teacher.getId().equals(id));
     }
-
 
     public void createTeacher(Teacher teacher) {
         List<Teacher> teacherList = teacherRepository.findAll();
